Fix simple replier, make changes to support asyncio with request reply, changes to support Python 2 and 3 in the same package
